requirements.txt
arabic-reshaper==2.1.4
asgiref==3.5.2
Django==3.2.9
django-dotenv==1.4.2
env==0.1.0
future==0.18.2
html5lib==1.1
Pillow==9.2.0
PyPDF2==2.11.1
python-bidi==0.4.2
pytz==2022.5
qrcode==7.3.1
reportlab==3.2.0
six==1.16.0
sqlparse==0.4.3
typing-extensions==4.4.0
webencodings==0.5.1
xhtml2pdf==0.2.5
